fix: send json headers in get requests

+++ packages/protocol/lib/src/messages/get_agent_status/get_agent_status_response.dart
@@ -18,8 +18,8 @@ class AgentStatusResponse {
   final String name;
   final String faction;
   final int numberOfShips;
-  final double cash;
-  final double totalAssets;
+  final int cash;
+  final int totalAssets;
   final bool gateOpen;
 
   Map<String, dynamic> toJson() => _$AgentStatusResponseToJson(this);
